[frontend] reorganize files contact services
